refactor(cli): extract enrichment/scoring cluster into server-enrichment.ts
First pass at breaking up the 3700-line server.ts (#354). Moves the cohesive,
pure source-enrichment and scan-prioritization cluster into a dedicated module:

- sourceSessionKey, pickSourceRecordForSession, selectCursorEnrichmentCandidates
- prioritizeScanInputs, mergeEnrichmentHints, enrichmentHintsFromBody
- the EnrichmentHints type
- internal helpers (preferenceScore, *PriorityScore, isRecentActivity,
  uniqueStrings, stringArrayFromUnknown, looksLikeCursorDisplayNoise)

server.ts imports the public symbols back, so __testables and all call sites are
unchanged. SourceSummaryRecord is referenced via a type-only import, so there is
no runtime import cycle. server.ts: 3761 -> 3562 lines; new module is 220.

No behavior change — all 340 CLI tests pass, typecheck and lint clean.

Refs #354
